+++ src/app/containers/Search/components.tsx
@@ -6,6 +6,5 @@ export const OuterContainer = styled.div`
   justify-content: center;
   align-items: center;
   height: 100vh;
-  border-right: 2px solid #444;
-  border-bottom: 2px solid #444;
+  box-shadow: inset 0px 0px 10px rgba(0, 0, 0, 0.1);
 `;