+++ src/app/containers/Preference/index.tsx
@@ -246,7 +246,6 @@ export default function PreferenceWindow() {
       setMacSystemPaths(),
     ])
       .then(() => {
-        console.log('resource initialized successfully.');
         return null;
       })
       .catch(console.error)